requirements.txt
astroid==2.4.1
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
pylint==2.5.2
python-telegram-bot==12.7
requests==2.23.0
schedule==0.6.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
tornado==6.0.4
urllib3==1.25.9
wrapt==1.12.1
